feature: job支持容器化部署 #83

+++ support-files/sql/job-analysis/0003_job_analysis_20210114-2000_V3.2.1.1_mysql.sql
@@ -1,6 +1,36 @@
 SET NAMES utf8mb4;
 USE job_analysis;
 
-ALTER TABLE `statistics` DROP INDEX `idx_appId_res_dim_dimValue`;
+DROP PROCEDURE IF EXISTS job_schema_update;
 
-ALTER TABLE `statistics` ADD INDEX `idx_appId_res_dim_dimValue_date`(`app_id`,`resource`(32), `dimension`(32), `dimension_value`(32),`date`(32)) USING BTREE;
+DELIMITER <JOB_UBF>
+
+CREATE PROCEDURE job_schema_update()
+BEGIN
+
+    DECLARE db VARCHAR(100);
+    SET AUTOCOMMIT = 0;
+    SELECT DATABASE() INTO db;
+
+    IF EXISTS(SELECT 1
+                  FROM information_schema.statistics
+                  WHERE TABLE_SCHEMA = db
+                    AND TABLE_NAME = 'statistics'
+                    AND INDEX_NAME = 'idx_appId_res_dim_dimValue') THEN
+        ALTER TABLE `statistics` DROP INDEX `idx_appId_res_dim_dimValue`;
+    END IF;
+
+    IF NOT EXISTS(SELECT 1
+                  FROM information_schema.statistics
+                  WHERE TABLE_SCHEMA = db
+                    AND TABLE_NAME = 'statistics'
+                    AND INDEX_NAME = 'idx_appId_res_dim_dimValue_date') THEN
+        ALTER TABLE `statistics` ADD INDEX `idx_appId_res_dim_dimValue_date`(`app_id`,`resource`(32), `dimension`(32), `dimension_value`(32),`date`(32)) USING BTREE;
+    END IF;
+
+COMMIT;
+END <JOB_UBF>
+DELIMITER ;
+CALL job_schema_update();
+
+DROP PROCEDURE IF EXISTS job_schema_update;
